Refactored new command into own class.
